Add type safe behavior creation to AnyMessage type
Motivation:
With upcoming update to behavior composition, we can remove
boiler plate around behvior casting `AnyMessage` type to the
type it handles.
AnyMessage type is extended with `be`(havior) function that does
a runtime check of the message type and returns `.unhandled` if
it's of a different type (it's not `Self`) than the type conforming
to `AnyMessage`.

Modifications:
- Extended `AnyMessage` protocol with behavior make function.

Result:
- Removed boiler plate around message casting in behavior implementation.